+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Tidarr notable changes.
 ### 🐛 Fixed
 - [API] Fix ignored umask issue
 - [API] Fix condition race issue with token refresh
+### 🖍 Changed
+- [API] Update to tiddl 3.1.1
+- [Front] Re-activate favorites download
+- [Front] Video download is back
 
 ## 📦 1.0.3
 ### 🚀 Added